Crate compress

dox (placeholder)

Re-exports

pub use self::checksum::adler::State32 as Adler32;

Modules

bwt
BWT (Burrows-Wheeler Transform) forward and backward transformation. Requires bwt feature, enabled by default

checksum
Checksum algorithms. Requires checksum feature, enabled by default

entropy
Entropy coder family. Requires entropy feature, enabled by default

flate
DEFLATE Compression and Decompression. Requires flate feature, enabled by default

lz4
LZ4 Decompression and Compression. Requires lz4 feature, enabled by default

rle
Run time length encoding and decoding based on byte streams, see https://en.wikipedia.org/wiki/Run-length_encoding.

zlib
ZLIB Compression and Decompression. Requires zlib feature, enabled by default

Traits

ReadExact
Adds a convenience method for types with the read trait, very similar to push_at_least in the late Reader trait