Function compress::bwt::encode_simple

pub fn encode_simple(input: &[Symbol]) -> (Vec<Symbol>, uint)

Transform an input block into the output slice, all-inclusive version. Returns the index of the original string in the output matrix.