Module compress::entropy::ari

Arithmetic encoder/decoder using the Range encoder underneath. Can be used in a general case of entropy coding stage. Supposed to be fast.

Links

http://en.wikipedia.org/wiki/Arithmetic_coding http://en.wikipedia.org/wiki/Range_encoding

Example

use std::io::{MemWriter, MemReader};
use compress::entropy::ari;

// Encode some text
let text = "some text";
let mut e = ari::ByteEncoder::new(MemWriter::new());
e.write_str(text);
let (encoded, _) = e.finish();

// Decode the encoded text
let mut d = ari::ByteDecoder::new(MemReader::new(encoded.unwrap()));
let decoded = d.read_to_end().unwrap();

Credit

This is an original implementation.

Reexports

pub use self::table::{ByteDecoder, ByteEncoder};

Modules

apm
Adaptive Probability Models

bin
Binary models for the arithmetic coder. The simplicity of the domain allows for normalized updates in place using bit shifts.

table
Frequency table models for the arithmetic coder. The module also implements Reader/Writer using simple byte coding.

Structs

Decoder
An arithmetic decoder helper

Encoder
An arithmetic encoder helper

RangeEncoder
Range Encoder basic primitive Gets probability ranges on the input, produces whole bytes of code on the output, where the code is an arbitrary fixed-ppoint value inside the resulting probability range.

Statics

pub static RANGE_DEFAULT_THRESHOLD: Border = 1<<14

Traits

Model
An abstract model to produce probability ranges Can be a table, a mix of tables, or just a smart function.

Type Definitions

Border

Symbol