Module compress::entropy::ari::apm

Adaptive Probability Models

Links

http://mattmahoney.net/dc/bbb.cpp

https://github.com/IlyaGrebnov/libbsc

Example

Credit

Matt Mahoney for the wonderful 'bbb' commented source

Structs

Bit
Bit probability model

Gate
Binary context gate maps an input binary probability into a new one by interpolating between internal maps in non-linear space

Type Definitions

BinCoords

FlatProbability

WideProbability