Struct compress::entropy::ari::apm::Gate

pub struct Gate {
    // some fields omitted
}

Binary context gate maps an input binary probability into a new one by interpolating between internal maps in non-linear space

Methods

impl Gate

fn new() -> Gate

Create a new gate instance

fn pass(&self, bit: &Bit) -> (Bit, BinCoords)

Pass a bit through the gate

fn pass_wide(&self, wp: WideProbability) -> (FlatProbability, BinCoords)

Pass a wide probability on input, usable when you mix it linearly beforehand (libbsc does that)

fn update_zero(&mut self, bc: BinCoords, rate: int, bias: int)

Mutate for better zeroes

fn update_one(&mut self, bc: BinCoords, rate: int, bias: int)

Mutate for better ones

fn update(&mut self, value: bool, bc: BinCoords, rate: int, bias: int)

Mutate for a given value