Struct compress::entropy::ari::bin::SumProxy

pub struct SumProxy<'a> {
    // some fields omitted
}

A proxy model for the combination of two binary models using equation: (wa * A + wb * B) >> ws

Methods

impl<'a> SumProxy<'a>

fn new(wa: Border, first: &'a Model, wb: Border, second: &'a Model, shift: Border) -> SumProxy<'a>

Create a new instance of the binary sum proxy