Struct compress::entropy::ari::Decoder

pub struct Decoder<R> {
    // some fields omitted
}

An arithmetic decoder helper

Methods

impl<R: Reader> Decoder<R>

fn new(r: R) -> Decoder<R>

Create a decoder on top of a given Reader

fn decode<V: Copy + Show, M: Model<V>>(&mut self, model: &M) -> IoResult<V>

Decode an abstract value based on the given Model

fn finish(self) -> (R, IoResult<()>)

Finish decoding