Struct compress::entropy::ari::RangeEncoder

pub struct RangeEncoder {
    pub threshold: Border,
    // some fields omitted
}

Range Encoder basic primitive Gets probability ranges on the input, produces whole bytes of code on the output, where the code is an arbitrary fixed-ppoint value inside the resulting probability range.

Fields

threshold

The minimum distance between low and hai to keep at all times, has to be at least the largest incoming 'total', and optimally many times larger

Methods

impl RangeEncoder

fn new(max_range: Border) -> RangeEncoder

Create a new instance will keep the active range below 'max_range'

fn reset(&mut self)

Reset the current range

fn process(&mut self, total: Border, from: Border, to: Border, output: &mut [Symbol]) -> uint

Process a given interval [from/total,to/total) into the current range write into the output slice, and return the number of symbols produced

fn query(&self, total: Border, code: Border) -> Border

Query the value encoded by 'code' in range [0,total)

fn get_code_tail(&mut self) -> Border

Get the code tail and close the range used at the end of encoding