Module compress::entropy::ari::table

Frequency table models for the arithmetic coder. The module also implements Reader/Writer using simple byte coding.

Links

Example

Credit

Structs

ByteDecoder
A basic byte-decoding arithmetic expects a special terminator code for the end of the stream

ByteEncoder
A basic byte-encoding arithmetic uses a special terminator code to end the stream

Model
A simple table of frequencies.

SumProxy
A proxy model for the sum of two frequency tables using equation: (wa * A + wb * B) >> ws

Type Definitions

Frequency