Struct compress::entropy::ari::table::Model

pub struct Model {
    // some fields omitted
}

A simple table of frequencies.

Methods

impl Model

fn new_custom(num_values: uint, threshold: Border, fn_init: |uint| -> Frequency) -> Model

Create a new table with frequencies initialized by a function

fn new_flat(num_values: uint, threshold: Border) -> Model

Create a new tanle with all frequencies being equal

fn reset_flat(&mut self)

Reset the table to the flat state

fn update(&mut self, value: uint, add_log: uint, add_const: Border)

Adapt the table in favor of given 'value' using 'add_log' and 'add_const' to produce the additive factor the higher 'add_log' is, the more concervative is the adaptation

fn downscale(&mut self)

Reduce frequencies by 'cut_iter' bits

fn get_frequencies<'a>(&'a self) -> &'a [Frequency]

Return read-only frequencies slice