Module compress::flate

DEFLATE Compression and Decompression

This module contains an implementation of the DEFLATE compression scheme. This format is often used as the underpinning of other compression formats.

Example

use compress::flate;
use std::io::File;

let stream = File::open(&Path::new("path/to/file.flate"));
let decompressed = flate::Decoder::new(stream).read_to_end();

Related links

http://tools.ietf.org/html/rfc1951 - RFC that this implementation is based on

http://www.gzip.org/zlib/rfc-deflate.html - simplified version of RFC 1951 used as a reference

http://svn.ghostscript.com/ghostscript/trunk/gs/zlib/contrib/puff/puff.c - Much of this code is based on the puff.c implementation found here

Structs

Decoder
The structure that is used to decode an LZ4 data stream. This wraps an internal reader which is used as the source of all data.