Struct compress::lz4::Decoder

pub struct Decoder<R> {
    pub r: R,
    // some fields omitted
}

This structure is used to decode a stream of LZ4 blocks. This wraps an internal reader which is read from when this decoder's read method is called.

Fields

r

The internally wrapped reader. This is exposed so it may be moved out of. Note that if data is read from the reader while decoding is in progress the output stream will get corrupted.

Methods

impl<R: Reader> Decoder<R>

fn new(r: R) -> Decoder<R>

Creates a new decoder which will read data from the given stream. The inner stream can be re-acquired by moving out of the r field of this structure.

fn reset(&mut self)

Resets this decoder back to its initial state. Note that the underlying stream is not seeked on or has any alterations performed on it.

fn eof(&mut self) -> bool

Tests whether the end of this LZ4 stream has been reached