Struct compress::lz4::Encoder

pub struct Encoder<W> {
    // some fields omitted
}

This structure is used to compress a stream of bytes using the LZ4 compression algorithm. This is a wrapper around an internal writer which bytes will be written to.

Methods

impl<W: Writer> Encoder<W>

fn new(w: W) -> Encoder<W>

Creates a new encoder which will have its output written to the given output stream. The output stream can be re-acquired by calling finish()

NOTE: compression isn't actually implemented just yet, this is just a skeleton of a future implementation.

fn finish(self) -> (W, IoResult<()>)

This function is used to flag that this session of compression is done with. The stream is finished up (final bytes are written), and then the wrapped writer is returned.