Module compress::bwt::mtf

MTF (Move To Front) encoder/decoder Produces a rank for each input character based on when it was seen last time. Useful for BWT output encoding, which produces a lot of zeroes and low ranks.

Links

http://en.wikipedia.org/wiki/Move-to-front_transform

Example

use std::io;
use compress::bwt::mtf;

// Encode a stream of bytes
let bytes = b"abracadabra";
let mut e = mtf::Encoder::new(io::MemWriter::new());
e.write(bytes).unwrap();
let encoded = e.finish().unwrap();

// Decode a stream of ranks
let mut d = mtf::Decoder::new(io::BufReader::new(encoded.as_slice()));
let decoded = d.read_to_end().unwrap();

Credit

Structs

Decoder
A simple MTF stream decoder

Encoder
A simple MTF stream encoder

MTF
MoveToFront encoder/decoder

Statics

pub static TOTAL_SYMBOLS: uint = 0x100

Type Definitions

Rank

Symbol